+++ src/streamtree/core/__init__.py
@@ -1,6 +1,7 @@
 """Core primitives: elements, components, render."""
 
 from streamtree.core.component import component, render, render_app
+from streamtree.core.context import debug_render_path
 from streamtree.core.element import ComponentCall, Element, Fragment, fragment
 
 __all__ = [
@@ -8,6 +9,7 @@ __all__ = [
     "Element",
     "Fragment",
     "component",
+    "debug_render_path",
     "fragment",
     "render",
     "render_app",